refactor: hook -> hooks 로 네이밍 통일
